Move withholding periods query logic to separate action
